shared core string builder
