+++ encoder/src/main/java/com/iabtcf/v2/encoder/ValueChecker.java
@@ -1,0 +1,10 @@
+package com.iabtcf.v2.encoder;
+
+/**
+ * @author ewhite 4/8/20
+ */
+@FunctionalInterface
+public interface ValueChecker {
+
+	boolean isSet(int value);
+}
